chore: fix clippy 1.95 lints
Address 82 clippy 1.95 errors across the H.264 crate by rewriting the
flagged code idiomatically where feasible and applying per-call-site
`#[allow]` with spec citations only where the indexed-loop pattern
mirrors an ITU-T H.264 spec equation verbatim. No crate-wide allow
shotgun: each suppression cites the exact subclause it preserves.

Highlights:
* `manual_checked_ops` → `checked_div` (5x macroblock_layer + 1x conformance test).
* `field_reassign_with_default` → struct-update syntax in 6 call sites.
* `if_same_then_else` → consolidated arms in poc.rs and macroblock_layer
  CAVLC chroma 4:4:4 path; the residual-derivation arms differ only by
  unused flag.
* `unnecessary_unwrap` → `if let Some(...)` in CABAC residual dispatch.
* `unnecessary_sort_by` → `sort_by_key(|e| Reverse(e.0))` for desc sorts.
* `vec_init_then_push` → `vec![...]` in scaling_list / vui tests.
* `needless_range_loop` rewritten with `iter_mut().enumerate()` where the
  loop variable was incidental; kept (with `// spec §X.Y.Z` comments) on
  raster-Z 4x4 / 8x8 walks (§8.5.10, §8.5.12, §8.5.13, §8.3.1.1, §8.3.2),
  spec eq. (8-18) FMO dispersed map, and Intra_8x8 ref-sample filter eq.
  (8-87).
* `manual_memcpy` → `copy_from_slice` for the AC-shift transform fixture.
* `drop_non_drop` on `drop(entropy)` kept with explanatory allow — the
  EntropyState owns mutable reborrows whose lifetimes we want to end
  before the I_PCM-recover branch reuses them.
* `large_enum_variant` on `decoder::Event` kept with allow — boxing
  `header`/`sps` fields would force every consumer to deref through
  `Box<…>` when destructuring `Event::Slice { … }`.
* `ptr_arg` (`&mut Vec<DpbEntry>` → `&mut [DpbEntry]`) on
  `apply_mmco` and `perform_marking`; neither pushed/popped.
* `manual_clamp`, `only_used_in_recursion`, `type_complexity`,
  `doc_lazy_continuation`, `doc_overindented_list_items` all rewritten
  idiomatically.

+++ src/inter_pred.rs
@@ -1106,10 +1106,10 @@ mod tests {
              imply 2x2 chroma for 4:2:0)"
         );
         // Integer-position copy: each output equals src at the given int.
-        assert_eq!(dst[0], ((4 * 2 + 4) & 0xff) as i32);
-        assert_eq!(dst[1], ((5 * 2 + 4) & 0xff) as i32);
-        assert_eq!(dst[2], ((4 * 2 + 5) & 0xff) as i32);
-        assert_eq!(dst[3], ((5 * 2 + 5) & 0xff) as i32);
+        assert_eq!(dst[0], ((4 * 2 + 4) & 0xff));
+        assert_eq!(dst[1], ((5 * 2 + 4) & 0xff));
+        assert_eq!(dst[2], ((4 * 2 + 5) & 0xff));
+        assert_eq!(dst[3], ((5 * 2 + 5) & 0xff));
     }
 
     #[test]
